Add files via upload
Flipkart is one of the leading e-commerce companies/platforms based in India.
The sentiment analysis of the product reviews is a way that can open ways to understand the performance of the products, which is beneficial for both future customers and the e-commerce platform.

The dataset I am using here for Flipkart reviews sentiment analysis is downloaded from Kaggle.

Conclusion:
By using this program we get the output, there are 923.55 Positive reviews, 96.77 Negative reviews, and 1283.68 Neutral reviews. So, most people give Neutral reviews, and a small proportion of people give Negative reviews. So we can say that people are satisfied with Flipkart products and services.